fix(web): align tag combobox interactions
